extractor/requirements.txt
# API and HTTP requests
requests==2.31.0
httpx==0.25.2

# Database
duckdb==0.9.2

# Data processing
pandas==2.1.4
numpy==1.24.3

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging and monitoring
structlog==23.2.0
rich==13.7.0

# Scheduling
schedule==1.2.0 